Update form handling code to use the complex fields
In the forms pull the list of complex fields to display out of the
database and use them to diplay and store the values.

Also update the PersonExtra code to use the complex fields model for the
getter that handles them. At the moment this is a bit sub optimal as it
requires a DB query to check if the field exists. I suspect we should
cache the relevant list when an instance of PersonExtra is created and
then check agains the cache.